Переделал логику сайта, теперь резкльтат поиска выводится в виде заголовков, подробное описание при переходе на нее, также добавил страницу со скилами, можно по скилам посмотреть в каких вакансиях они есть

+++ HHparser/parserapp/parser.py
@@ -4,13 +4,16 @@ import time
 from parserapp.models import Vacancy, Skills_table, Params
 
 
-def hh_serch(tex, param):
+def hh_serch(tex, param, del_bd):
 
     url = 'https://api.hh.ru/vacancies'
 
-    Vacancy.objects.all().delete()
     Params.objects.all().delete()
-    Skills_table.objects.all().delete()
+    if del_bd == 'delit':
+        Vacancy.objects.all().delete()
+        Skills_table.objects.all().delete()
+    else:
+        pass
 
     if param == "name":
         ser = 'В названии вакансии '
@@ -53,17 +56,20 @@ def hh_serch(tex, param):
                 else:
                     zp = 'Не указана'
 
+                ab = result['items'][z]['snippet']['responsibility'].replace('<highlighttext>', '')
+                ab = ab.replace('</highlighttext>','')
+
                 vac = Vacancy.objects.create(name=result['items'][z]['name'], salary=zp,
-                                       about=result['items'][z]['snippet']['responsibility'],
+                                       about=ab,
                                        link=result['items'][z]['alternate_url'])
 
                 for skills_vac in requests.get(result['items'][z]['url']).json()['key_skills']:
                     skill_vacancy = skills_vac['name']
-                    if skill_vacancy not in skills:
-
-                        skills.append(skill_vacancy)
+                    # if skill_vacancy not in skills:
+                    #
+                    #     skills.append(skill_vacancy)
                         # try:
-                        vac.skils.create(skil=skill_vacancy)
+                    vac.skils.create(skil=skill_vacancy)
                         # except:
                         #     pass
